requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.7
Django==5.0.6
django-allauth==0.63.6
django-bootstrap5==24.2
django-celery-beat==2.6.0
django-timezone-field==7.0
djangorestframework==3.15.2
gcloud==0.18.3
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
jwcrypto==1.5.6
kombu==5.3.7
oauth2client==4.1.3
prompt_toolkit==3.0.47
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pyparsing==3.1.2
Pyrebase4==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-env==1.0.0
python-jwt==4.1.0
requests==2.32.2
requests-toolbelt==0.10.1
rsa==4.9
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
